requirements.txt
django
gunicorn
django-heroku
requests
pyGithub